refactor(analysis): externalize solver inline tests
