+++ foundation/analysis/src/linalg/solver/lu.rs
@@ -200,96 +200,3 @@ impl<T: Scalar> LinearSolver<T> for LUSolver<T> {
         Ok(SolutionInfo::direct(solution, residual))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_lu_decomposition_2x2() {
-        let matrix = vec![vec![2.0, 1.0], vec![1.0, 3.0]];
-
-        let solver = LUSolver::new(1e-15);
-        let lu_decomp = solver.decompose(&matrix).unwrap();
-
-        // LU分解の検証（PA = LU）
-        let n = matrix.len();
-        for i in 0..n {
-            for (j, _) in matrix.iter().enumerate().take(n) {
-                let mut reconstructed = 0.0;
-                for k in 0..n {
-                    let l_ik = if i > k {
-                        lu_decomp.lu_matrix[i][k]
-                    } else if i == k {
-                        1.0
-                    } else {
-                        0.0
-                    };
-                    let u_kj = if k <= j {
-                        lu_decomp.lu_matrix[k][j]
-                    } else {
-                        0.0
-                    };
-                    reconstructed += l_ik * u_kj;
-                }
-                let original = matrix[lu_decomp.permutation[i]][j];
-                assert!((reconstructed - original).abs() < 1e-10);
-            }
-        }
-    }
-
-    #[test]
-    fn test_lu_solver_2x2() {
-        let matrix = vec![vec![2.0, 1.0], vec![1.0, 3.0]];
-        let rhs = vec![5.0, 6.0];
-
-        let solver = LUSolver::new(1e-15);
-        let result = solver.solve(&matrix, &rhs).unwrap();
-
-        assert!((result.solution[0] - 1.8).abs() < 1e-10);
-        assert!((result.solution[1] - 1.4).abs() < 1e-10);
-        assert!(result.converged);
-    }
-
-    #[test]
-    fn test_lu_solver_3x3() {
-        let matrix = vec![
-            vec![2.0, 1.0, -1.0],
-            vec![-3.0, -1.0, 2.0],
-            vec![-2.0, 1.0, 2.0],
-        ];
-        let rhs = vec![8.0, -11.0, -3.0];
-
-        let solver = LUSolver::new(1e-15);
-        let result = solver.solve(&matrix, &rhs).unwrap();
-
-        assert!((result.solution[0] - 2.0).abs() < 1e-10);
-        assert!((result.solution[1] - 3.0).abs() < 1e-10);
-        assert!((result.solution[2] - (-1.0)).abs() < 1e-10);
-    }
-
-    #[test]
-    fn test_lu_determinant() {
-        let matrix = vec![vec![2.0, 1.0], vec![1.0, 3.0]];
-
-        let solver = LUSolver::new(1e-15);
-        let lu_decomp = solver.decompose(&matrix).unwrap();
-        let det = solver.determinant(&lu_decomp);
-
-        // 行列式 = 2*3 - 1*1 = 5
-        assert!((det - 5.0).abs() < 1e-10);
-    }
-
-    #[test]
-    fn test_lu_singular_matrix() {
-        let matrix = vec![
-            vec![1.0, 2.0],
-            vec![2.0, 4.0], // 2倍の行
-        ];
-
-        let solver = LUSolver::new(1e-15);
-        let result = solver.decompose(&matrix);
-
-        assert!(result.is_err());
-    }
-}
