+++ foundation/analysis/src/linalg/solver/newton.rs
@@ -174,76 +174,3 @@ where
 
     None
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_newton_solve_square_root() {
-        // x^2 - 2 = 0 の解を求める(√2を計算)
-        let f = |x: f64| x * x - 2.0;
-        let df = |x: f64| 2.0 * x;
-        let result = newton_solve(f, df, 1.0, 100, 1e-10);
-
-        assert!(result.is_some());
-        let sqrt_2 = result.unwrap();
-        assert!((sqrt_2 - std::f64::consts::SQRT_2).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_newton_inverse_cube_root() {
-        // x^3 の逆関数（立方根）を計算
-        let f = |x: f64| x * x * x;
-        let df = |x: f64| 3.0 * x * x;
-        let result = newton_inverse(f, df, 8.0, 2.0, 100, 1e-10);
-
-        assert!(result.is_some());
-        let cube_root = result.unwrap();
-        assert!((cube_root - 2.0).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_newton_solve_zero_derivative() {
-        // 導関数が0になるケース
-        let f = |x: f64| x * x;
-        let df = |_: f64| 0.0; // 常に0の導関数
-        let result = newton_solve(f, df, 1.0, 100, 1e-10);
-
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_newton_solve_2d_circle_line() {
-        // 連立方程式: x^2 + y^2 = 1, x - y = 0
-        // 単位円と y=x の交点を求める
-        let system = |x: f64, y: f64| {
-            let f1 = x * x + y * y - 1.0;
-            let f2 = x - y;
-            let jacobian = [[2.0 * x, 2.0 * y], [1.0, -1.0]];
-            (f1, f2, jacobian)
-        };
-
-        let result = newton_solve_2d(system, (1.0, 0.5), 100, 1e-10);
-        assert!(result.is_some());
-
-        let (x, y) = result.unwrap();
-        let expected = 1.0 / 2_f64.sqrt();
-        assert!((x - expected).abs() < 1e-6);
-        assert!((y - expected).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_newton_solve_2d_singular_jacobian() {
-        // ヤコビ行列が特異になるケース
-        let system = |x: f64, y: f64| {
-            let f1 = x + y;
-            let f2 = x + y; // f2 = f1（従属）
-            let jacobian = [[1.0, 1.0], [1.0, 1.0]]; // det = 0
-            (f1, f2, jacobian)
-        };
-
-        let result = newton_solve_2d(system, (1.0, 1.0), 100, 1e-10);
-        assert!(result.is_none());
-    }
-}
