+++ foundation/analysis/src/linalg/solver/newton_tests.rs
@@ -1,0 +1,68 @@
+use super::newton::{newton_inverse, newton_solve, newton_solve_2d};
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_newton_solve_square_root() {
+        let f = |x: f64| x * x - 2.0;
+        let df = |x: f64| 2.0 * x;
+        let result = newton_solve(f, df, 1.0, 100, 1e-10);
+
+        assert!(result.is_some());
+        let sqrt_2 = result.unwrap();
+        assert!((sqrt_2 - std::f64::consts::SQRT_2).abs() < 1e-6);
+    }
+
+    #[test]
+    fn test_newton_inverse_cube_root() {
+        let f = |x: f64| x * x * x;
+        let df = |x: f64| 3.0 * x * x;
+        let result = newton_inverse(f, df, 8.0, 2.0, 100, 1e-10);
+
+        assert!(result.is_some());
+        let cube_root = result.unwrap();
+        assert!((cube_root - 2.0).abs() < 1e-6);
+    }
+
+    #[test]
+    fn test_newton_solve_zero_derivative() {
+        let f = |x: f64| x * x;
+        let df = |_: f64| 0.0;
+        let result = newton_solve(f, df, 1.0, 100, 1e-10);
+
+        assert!(result.is_none());
+    }
+
+    #[test]
+    fn test_newton_solve_2d_circle_line() {
+        let system = |x: f64, y: f64| {
+            let f1 = x * x + y * y - 1.0;
+            let f2 = x - y;
+            let jacobian = [[2.0 * x, 2.0 * y], [1.0, -1.0]];
+            (f1, f2, jacobian)
+        };
+
+        let result = newton_solve_2d(system, (1.0, 0.5), 100, 1e-10);
+        assert!(result.is_some());
+
+        let (x, y) = result.unwrap();
+        let expected = 1.0 / 2_f64.sqrt();
+        assert!((x - expected).abs() < 1e-6);
+        assert!((y - expected).abs() < 1e-6);
+    }
+
+    #[test]
+    fn test_newton_solve_2d_singular_jacobian() {
+        let system = |x: f64, y: f64| {
+            let f1 = x + y;
+            let f2 = x + y;
+            let jacobian = [[1.0, 1.0], [1.0, 1.0]];
+            (f1, f2, jacobian)
+        };
+
+        let result = newton_solve_2d(system, (1.0, 1.0), 100, 1e-10);
+        assert!(result.is_none());
+    }
+}
